Detached Schedules into SchedulesMixin
