+++ gvm/protocols/gmpv208/gmpv208.py
@@ -466,133 +466,6 @@ class GmpV208Mixin(GvmProtocol):
 
         return self._send_xml_command(cmd)
 
-    def create_schedule(
-        self,
-        name: str,
-        icalendar: str,
-        timezone: str,
-        *,
-        comment: Optional[str] = None,
-    ) -> Any:
-        """Create a new schedule based in `iCalendar`_ data.
-
-        Example:
-            Requires https://pypi.org/project/icalendar/
-
-            .. code-block:: python
-
-                import pytz
-
-                from datetime import datetime
-
-                from icalendar import Calendar, Event
-
-                cal = Calendar()
-
-                cal.add('prodid', '-//Foo Bar//')
-                cal.add('version', '2.0')
-
-                event = Event()
-                event.add('dtstamp', datetime.now(tz=pytz.UTC))
-                event.add('dtstart', datetime(2020, 1, 1, tzinfo=pytz.utc))
-
-                cal.add_component(event)
-
-                gmp.create_schedule(
-                    name="My Schedule",
-                    icalendar=cal.to_ical(),
-                    timezone='UTC'
-                )
-        Arguments:
-            name: Name of the new schedule
-            icalendar: `iCalendar`_ (RFC 5545) based data.
-            timezone: Timezone to use for the icalender events e.g
-                Europe/Berlin. If the datetime values in the icalendar data are
-                missing timezone information this timezone gets applied.
-                Otherwise the datetime values from the icalendar data are
-                displayed in this timezone
-            comment: Comment on schedule.
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-
-        .. _iCalendar:
-            https://tools.ietf.org/html/rfc5545
-        """
-        if not name:
-            raise RequiredArgument(
-                function=self.create_schedule.__name__, argument='name'
-            )
-        if not icalendar:
-            raise RequiredArgument(
-                function=self.create_schedule.__name__, argument='icalendar'
-            )
-        if not timezone:
-            raise RequiredArgument(
-                function=self.create_schedule.__name__, argument='timezone'
-            )
-
-        cmd = XmlCommand("create_schedule")
-
-        cmd.add_element("name", name)
-        cmd.add_element("icalendar", icalendar)
-        cmd.add_element("timezone", timezone)
-
-        if comment:
-            cmd.add_element("comment", comment)
-
-        return self._send_xml_command(cmd)
-
-    def modify_schedule(
-        self,
-        schedule_id: str,
-        *,
-        name: Optional[str] = None,
-        icalendar: Optional[str] = None,
-        timezone: Optional[str] = None,
-        comment: Optional[str] = None,
-    ) -> Any:
-        """Modifies an existing schedule
-
-        Arguments:
-            schedule_id: UUID of the schedule to be modified
-            name: Name of the schedule
-            icalendar: `iCalendar`_ (RFC 5545) based data.
-            timezone: Timezone to use for the icalender events e.g
-                Europe/Berlin. If the datetime values in the icalendar data are
-                missing timezone information this timezone gets applied.
-                Otherwise the datetime values from the icalendar data are
-                displayed in this timezone
-            commenhedule.
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-
-        .. _iCalendar:
-            https://tools.ietf.org/html/rfc5545
-        """
-        if not schedule_id:
-            raise RequiredArgument(
-                function=self.modify_schedule.__name__, argument='schedule_id'
-            )
-
-        cmd = XmlCommand("modify_schedule")
-        cmd.set_attribute("schedule_id", schedule_id)
-
-        if name:
-            cmd.add_element("name", name)
-
-        if icalendar:
-            cmd.add_element("icalendar", icalendar)
-
-        if timezone:
-            cmd.add_element("timezone", timezone)
-
-        if comment:
-            cmd.add_element("comment", comment)
-
-        return self._send_xml_command(cmd)
-
     def clone_filter(self, filter_id: str) -> Any:
         """Clone an existing filter
 
@@ -775,24 +648,6 @@ class GmpV208Mixin(GvmProtocol):
         cmd.add_element("copy", role_id)
         return self._send_xml_command(cmd)
 
-    def clone_schedule(self, schedule_id: str) -> Any:
-        """Clone an existing schedule
-
-        Arguments:
-            schedule_id: UUID of an existing schedule to clone from
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not schedule_id:
-            raise RequiredArgument(
-                function=self.clone_schedule.__name__, argument='schedule_id'
-            )
-
-        cmd = XmlCommand("create_schedule")
-        cmd.add_element("copy", schedule_id)
-        return self._send_xml_command(cmd)
-
     def delete_filter(
         self, filter_id: str, *, ultimate: Optional[bool] = False
     ) -> Any:
@@ -879,26 +734,6 @@ class GmpV208Mixin(GvmProtocol):
 
         cmd = XmlCommand("delete_role")
         cmd.set_attribute("role_id", role_id)
-        cmd.set_attribute("ultimate", to_bool(ultimate))
-
-        return self._send_xml_command(cmd)
-
-    def delete_schedule(
-        self, schedule_id: str, *, ultimate: Optional[bool] = False
-    ) -> Any:
-        """Deletes an existing schedule
-
-        Arguments:
-            schedule_id: UUID of the schedule to be deleted.
-            ultimate: Whether to remove entirely, or to the trashcan.
-        """
-        if not schedule_id:
-            raise RequiredArgument(
-                function=self.delete_schedule.__name__, argument='schedule_id'
-            )
-
-        cmd = XmlCommand("delete_schedule")
-        cmd.set_attribute("schedule_id", schedule_id)
         cmd.set_attribute("ultimate", to_bool(ultimate))
 
         return self._send_xml_command(cmd)
@@ -1198,63 +1033,6 @@ class GmpV208Mixin(GvmProtocol):
 
         cmd = XmlCommand("get_roles")
         cmd.set_attribute("role_id", role_id)
-        return self._send_xml_command(cmd)
-
-    def get_schedules(
-        self,
-        *,
-        filter: Optional[str] = None,
-        filter_id: Optional[str] = None,
-        trash: Optional[bool] = None,
-        tasks: Optional[bool] = None,
-    ) -> Any:
-        """Request a list of schedules
-
-        Arguments:
-            filter: Filter term to use for the query
-            filter_id: UUID of an existing filter to use for the query
-            trash: Whether to get the trashcan schedules instead
-            tasks: Whether to include tasks using the schedules
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        cmd = XmlCommand("get_schedules")
-
-        add_filter(cmd, filter, filter_id)
-
-        if trash is not None:
-            cmd.set_attribute("trash", to_bool(trash))
-
-        if tasks is not None:
-            cmd.set_attribute("tasks", to_bool(tasks))
-
-        return self._send_xml_command(cmd)
-
-    def get_schedule(
-        self, schedule_id: str, *, tasks: Optional[bool] = None
-    ) -> Any:
-        """Request a single schedule
-
-        Arguments:
-            schedule_id: UUID of an existing schedule
-            tasks: Whether to include tasks using the schedules
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        cmd = XmlCommand("get_schedules")
-
-        if not schedule_id:
-            raise RequiredArgument(
-                function=self.get_schedule.__name__, argument='schedule_id'
-            )
-
-        cmd.set_attribute("schedule_id", schedule_id)
-
-        if tasks is not None:
-            cmd.set_attribute("tasks", to_bool(tasks))
-
         return self._send_xml_command(cmd)
 
     def get_settings(self, *, filter: Optional[str] = None) -> Any:
